corrected one user can like only one time

+++ controllers/like.js
@@ -9,10 +9,18 @@ const addLike = (req, res) => {
     if (err) return res.status(403).json("Token is not valid!");
 
     const { post_id, user_id } = req.body;
-    const q = "INSERT INTO likes (post_id, user_id) VALUES (?, ?)";
-    db.query(q, [post_id, user_id], (err, result) => {
-      if (err) return res.status(500).send(err);
-      res.status(201).send("Like added");
+    const checkQuery = 'SELECT * FROM likes WHERE post_id = ? AND user_id = ?';
+    db.query(checkQuery, [post_id, user_id], (checkErr, checkResult) => {
+      if (checkErr) return res.status(500).send(checkErr);
+      if (checkResult.length > 0) {
+        return res.status(400).send('User has already liked this post');
+      } else {
+        const q = "INSERT INTO likes (post_id, user_id) VALUES (?, ?)";
+        db.query(q, [post_id, user_id], (insertErr, insertResult) => {
+          if (insertErr) return res.status(500).send(insertErr);
+          res.status(201).send("Like added");
+        });
+      }
     });
   });
 };
@@ -25,10 +33,18 @@ const deleteLike = (req, res) => {
     if (err) return res.status(403).json("Token is not valid!");
 
     const { post_id, user_id } = req.body;
-    const q = "DELETE FROM likes WHERE post_id = ? AND user_id = ?";
-    db.query(q, [post_id, user_id], (err, result) => {
-      if (err) return res.status(500).send(err);
-      res.status(200).send("Like removed");
+    const checkQuery = 'SELECT * FROM likes WHERE post_id = ? AND user_id = ?';
+    db.query(checkQuery, [post_id, user_id], (checkErr, checkResult) => {
+      if (checkErr) return res.status(500).send(checkErr);
+      if (checkResult.length === 0) {
+        return res.status(400).send('User has not liked this post');
+      } else {
+        const q = "DELETE FROM likes WHERE post_id = ? AND user_id = ?";
+        db.query(q, [post_id, user_id], (deleteErr, deleteResult) => {
+          if (deleteErr) return res.status(500).send(deleteErr);
+          res.status(200).send("Like removed");
+        });
+      }
     });
   });
 };
